# Conflicts:
#	models/favorite.go
#	router/routes.go

+++ models/follower.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type Follower struct {
 	gorm.Model
-	HostID     int64
-	FollowerID int64
+	HostID     uint
+	FollowerID uint
 }
